Move fetchCampaigns to helpers
