+++ lib/helpers.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const request = require('superagent');
+
 /**
  * Returns Gambit URI for production, or Thor if set as environmentName.
  */
@@ -22,4 +24,51 @@ module.exports.phoenixBaseUri = function (environmentName) {
   }
 
   return `https://${subdomain}dosomething.org/`;
+};
+
+/**
+ * Fetches Gambit campaigns on production, or for given environment.
+ */
+module.exports.fetchCampaigns = function (environmentName) {
+  const output = {
+    username: 'Puppet Sloth',
+    text: `Gambit Campaigns running on *${environmentName.toUpperCase()}*:`,
+    mrkdwn: true,
+    attachments: [],
+  };
+
+  const colors = ['FCD116', '23b7fb', '4e2b63'];
+
+  return new Promise((resolve, reject) => {
+    const gambitCampaignsUri = `${this.gambitApiBaseUri(environmentName)}campaigns`;
+
+    return request.get(gambitCampaignsUri)
+      .then((response) => {
+        response.body.data.forEach((campaign, index) => {
+          const campaignUri = `${this.phoenixBaseUri(environmentName)}node/${campaign.id}`;
+          const keywords = campaign.keywords.map(entry => entry.keyword);
+
+          output.attachments.push({
+            color: `#${colors[index % 3]}`,
+            title: `${campaign.title} (ID ${campaign.id})`,
+            title_link: campaignUri,
+            fields: [
+              {
+                title: 'Keywords',
+                value: keywords.join(', '),
+                short: true,
+              },
+              {
+                title: 'Status',
+                value: campaign.status,
+                short: true,
+              },
+            ],
+          });
+        });
+
+        return resolve(output);
+      })
+    .catch(err => reject(err));
+  });
 };
